requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
seaborn>=0.13.0
matplotlib>=3.8.0
openpyxl>=3.1.0
lxml>=4.9.0
html5lib>=1.1
